Use image that downloads Factorio, support running the game
Moves to using a single docker image that downloads Factorio
just-in-time based on the provided version or channel. This helps us
ensure that the base image itself is always up-to-date.

Added the logic to start the game as well as configure it for first-run.
Not super well tested, but will be refactoring this and adding tests
shortly.